점수 계산의 책임을 Hand가 가지도록 수정
